Add files via upload
Density-Based Spatial Clustering of Applications with Noise
is known as DBSCAN. It is a clustering algorithm based on density. The
algorithm turns regions with a high enough density into clusters and searches
spatial databases with noise for clusters with any possible architecture. It
depicts a cluster as the largest collection of densely connected points.